Release candidate 3 > Wyczyszczenie kodu ze zbednych smieci
